get rid of the extra, project-specific (python) package
in practice we found it hard to keep project-specific things in a separate package and leave `restbase` mostly untouched.  also, the location of existing code and the question where to put new code causes confusion rather quickly.

the new policy is to only use one package, which we will extend and modify to meet a project's needs.  in order to make it easier to merge things back and forth, we'll keep the module name generic, though.